+++ src/components/__tests__/Navigation.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Navigation from './Navigation';
+import Navigation from '../Navigation';
 
 describe('<Navigation />', () => {
